Rough version which is able to get plurks of a user without logging on.  A temporary scraper is used to find the user_id given a username.  The user_id is returned when you log on which is why it was required in order to read someone's plurks.
Class methods are now used to access the functionality which does not require someone to be logged in, which currently only is applicable to get_plurks.  An alias was added to name the method to plurks_for.  There is also a class method login which returns a Plurk::Base object upon successful creation.  Considering re-naming Plurk::Base to Plurk::User but that's a radicle change to discuss with others.